bug 1483246: extract graphics report into new django app
