+++ webapp-django/crashstats/crashstats/urls.py
@@ -48,9 +48,6 @@ urlpatterns = [
     url(r'^login/$',
         views.login,
         name='login'),
-    url(r'^graphics_report/$',
-        views.graphics_report,
-        name='graphics_report'),
     url(r'^about/throttling/$',
         views.about_throttling,
         name='about_throttling'),
